src/chat_app/requirements.txt
gradio>=4.0.0,<5.0.0
databricks-sdk>=0.30.0
